Move stuff around, simplify package.json
