+++ packages/site-kit/src/lib/components/ThemeToggle.svelte
@@ -1,7 +1,31 @@
-<script>
+<script context="module">
 	import { browser } from '$app/environment';
+	import { persisted } from 'svelte-local-storage-store';
+
+	/**
+	 * @typedef {{ preference: 'light' | 'dark' | 'system', current: 'light' | 'dark' }} Theme
+	 */
+
+	/** @type {import('svelte/store').Writable<Theme>} */
+	export const theme = persisted('svelte:theme', {
+		preference: 'system',
+		current: browser
+			? window.matchMedia('(prefers-color-scheme: dark)').matches
+				? 'dark'
+				: 'light'
+			: 'light'
+	});
+
+	theme.subscribe(($theme) => {
+		if (!browser) return;
+
+		document.body.classList.remove('light', 'dark');
+		document.body.classList.add($theme.current);
+	});
+</script>
+
+<script>
 	import { onDestroy } from 'svelte';
-	import { theme } from './store';
 
 	function toggle() {
 		const upcoming_theme = $theme.current === 'light' ? 'dark' : 'light';
