+++ packages/site-kit/src/lib/components/TryTerminal.svelte
