--- packages/site-kit/src/lib/components/home/index.js
@@ -1,1 +1,0 @@
-export { default as TryTerminal } from './TryTerminal.svelte';
