+++ packages/site-kit/src/lib/components/index.js
@@ -9,3 +9,4 @@ export { default as Separator } from './Separator.svelte';
 export { default as Shell } from './Shell.svelte';
 export { default as SkipLink } from './SkipLink.svelte';
 export { default as TSToggle } from './TSToggle.svelte';
+export { default as TryTerminal } from './TryTerminal.svelte';
