+++ packages/site-kit/src/lib/styles/tokens.css
@@ -215,49 +215,3 @@
 		--sk-code-diff-removed: hsl(2, 64%, 52%);
 	}
 }
-
-/* legacy — existing site-kit uses these, and upgrading svelte.dev will be easier if we keep them around for now */
-/* body {
-	--back: var(--sk-back-1);
-	--back-light: var(--sk-back-3);
-	--back-api: var(--sk-back-4);
-	--prime: var(--sk-theme-1);
-	--second: var(--sk-theme-2);
-	--flash: var(--sk-theme-3);
-	--heading: var(--sk-text-1);
-	--text: var(--sk-text-2);
-	--sidebar-text: rgba(255, 255, 255, 0.9);
-
-	--code-bg: var(--sk-code-bg);
-	--code-base: var(--sk-code-base);
-	--code-comment: var(--sk-code-comment);
-	--code-keyword: var(--sk-code-keyword);
-	--code-function: var(--sk-code-function);
-	--code-string: var(--sk-code-string);
-	--code-number: var(--sk-code-number);
-	--code-template-string: var(--sk-code-template-string);
-	--code-tags: var(--sk-code-tags);
-	--code-important: var(--sk-code-important);
-
-	--nav-h: var(--sk-nav-height);
-	--top-offset: var(--sk-page-padding-top);
-	--side-nav: var(--sk-page-padding-side);
-	--sidebar-w: var(--sidebar-width);
-	--border-w: var(--sk-thick-border-width);
-	--border-r: var(--sk-border-radius);
-
-	--font: var(--sk-font);
-	--font-mono: var(--sk-font-mono);
-	--h1: var(--sk-text-xxl);
-	--h2: var(--sk-text-l);
-	--h3: var(--sk-text-m);
-	--h4: var(--sk-text-s);
-	--h5: var(--sk-text-s);
-	--h6: var(--sk-text-xs);
-	--code-fs: var(--sk-text-xs);
-	--linemax: var(--sk-line-max-width);
-
-	--main-width: 80rem;
-	--sidebar-mid-w: 36rem;
-	--sidebar-large-w: 48rem;
-} */
